+++ tests/unit/Builders.FromPackageJson.ComponentBuilder.spec.js
@@ -18,28 +18,25 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) OWASP Foundation. All Rights Reserved.
 */
 
-module.exports = {
+const assert = require('assert')
+const { suite, test } = require('mocha')
 
-  /**
-   * Capitalise the first letter of a string
-   * @param {string} s
-   * @return {string}
-   */
-  capitaliseFirstLetter: s => s.charAt(0).toUpperCase() + s.slice(1),
-  /**
-   * UpperCamelCase a string
-   * @param {string} s
-   * @return {string}
-   */
-  upperCamelCase: s => s.replace(
-    /\b\w/g,
-    f => f.slice(-1).toUpperCase()
-  ).replace(/\W/g, ''),
+const {
+  Builders: { FromPackageJson: { ComponentBuilder } },
+  Factories: {
+    FromPackageJson: { ExternalReferenceFactory },
+    LicenseFactory
+  }
+} = require('../../')
 
-  /**
-   * Generate a random string of length.
-   * @param {number} length
-   * @return {string}
-   */
-  randomString: length => Math.random().toString(32).substring(2, 2 + length)
-}
+suite('Builders.FromPackageJson.ToolBuilder', () => {
+  test('construct', () => {
+    const extRefFactory = new ExternalReferenceFactory()
+    const licenseFactory = new LicenseFactory()
+
+    const actual = new ComponentBuilder(extRefFactory, licenseFactory)
+
+    assert.strictEqual(actual.extRefFactory, extRefFactory)
+    assert.strictEqual(actual.licenseFactory, licenseFactory)
+  })
+})