+++ src/builders/index.node.ts
@@ -17,4 +17,6 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) OWASP Foundation. All Rights Reserved.
 */
 
-export * as FromPackageJson from './fromPackageJson.node'
+export * as FromNodePackageJson from './fromNodePackageJson.node'
+/** @deprecated use {@see FromNodePackageJson} instead of {@see FromPackageJson} */
+export * as FromPackageJson from './fromNodePackageJson.node'