+++ tests/integration/Builders.FromNodePackageJson.ToolBuilders.test.js
@@ -22,21 +22,33 @@ const assert = require('assert')
 const { suite, test } = require('mocha')
 
 const {
-  Builders: { FromPackageJson: { ComponentBuilder } },
-  Factories: {
-    FromPackageJson: { ExternalReferenceFactory },
-    LicenseFactory
-  }
+  Models,
+  Factories,
+  Builders: { FromNodePackageJson: { ToolBuilder } }
 } = require('../../')
 
-suite('Builders.FromPackageJson.ToolBuilder', () => {
-  test('construct', () => {
-    const extRefFactory = new ExternalReferenceFactory()
-    const licenseFactory = new LicenseFactory()
+suite('Builders.FromNodePackageJson.ToolBuilder', () => {
+  const salt = Math.random()
 
-    const actual = new ComponentBuilder(extRefFactory, licenseFactory)
+  const extRefFactory = new Factories.FromNodePackageJson.ExternalReferenceFactory()
+  extRefFactory.makeExternalReferences = () => [`FAKE REFERENCES ${salt}`]
 
-    assert.strictEqual(actual.extRefFactory, extRefFactory)
-    assert.strictEqual(actual.licenseFactory, licenseFactory)
+  const sut = new ToolBuilder(extRefFactory)
+
+  const data = {
+    name: '@foo/bar',
+    version: `1.33.7-alpha.23.${salt}`
+    // to be continued
+  }
+  const expected = new Models.Tool({
+    vendor: '@foo',
+    name: 'bar',
+    version: `1.33.7-alpha.23.${salt}`,
+    externalReferences: new Models.ExternalReferenceRepository([`FAKE REFERENCES ${salt}`])
+  })
+
+  test('makeTool', () => {
+    const actual = sut.makeTool(data)
+    assert.deepStrictEqual(actual, expected)
   })
 })