+++ src/builders/index.node.ts
@@ -17,6 +17,7 @@ SPDX-License-Identifier: Apache-2.0
 Copyright (c) OWASP Foundation. All Rights Reserved.
 */
 
+/** @since 1.2.0 */
 export * as FromNodePackageJson from './fromNodePackageJson.node'
-/** @deprecated use {@see FromNodePackageJson} instead of {@see FromPackageJson} */
+/** @deprecated use {@link FromNodePackageJson} instead of {@link FromPackageJson} */
 export * as FromPackageJson from './fromNodePackageJson.node'