+++ src/factories/index.node.ts
@@ -19,6 +19,7 @@ Copyright (c) OWASP Foundation. All Rights Reserved.
 
 export * from './index.common'
 
+/** @since 1.2.0 */
 export * as FromNodePackageJson from './fromNodePackageJson.node'
-/** @deprecated use {@see FromNodePackageJson} instead of {@see FromPackageJson} */
+/** @deprecated use {@link FromNodePackageJson} instead of {@link FromPackageJson} */
 export * as FromPackageJson from './fromNodePackageJson.node'